feat: Add pytest test utils and unittest migration plan
Adds a new pytest-friendly test utility file at `test/test_utils_new.py`. This file includes the `capture_logs` context manager and decorator for advanced log testing, along with other test helpers.

Moves the `get_logger_children` function from the test utilities to `ibind/support/logs.py` to make it a general-purpose utility.

Creates a detailed migration plan in `test/migration_plan.md` to guide the conversion of all existing unittest files to pytest. The plan instructs developers to create new test files for side-by-side comparison and to use the new utilities.
