+++ ibind/support/logs.py
@@ -2,6 +2,7 @@ import datetime
 import logging
 import sys
 from pathlib import Path
+from typing import List
 
 from ibind import var
 
@@ -9,6 +10,25 @@ DEFAULT_FORMAT = '%(asctime)s|%(levelname)-.1s| %(message)s'
 
 _initialized = False
 _log_to_file = False
+
+
+def get_logger_children(main_logger) -> List[logging.Logger]:
+    """
+    Gets child loggers. Added as a support compat for Python version 3.11 and below.
+    Source: https://github.com/python/cpython/blob/3.12/Lib/logging/__init__.py#L1831
+    """
+    if hasattr(main_logger, 'getChildren'):
+        return list(main_logger.getChildren())
+
+    def _hierlevel(logger):
+        if logger is logger.manager.root:
+            return 0
+        return 1 + logger.name.count('.')
+
+    d = main_logger.manager.loggerDict
+    return [item for item in d.values()
+            if isinstance(item, logging.Logger) and item.parent is main_logger and
+            _hierlevel(item) == 1 + _hierlevel(item.parent)]
 
 
 def project_logger(filepath=None):
